% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeGPT.R
\name{CellTypeGPT}
\alias{CellTypeGPT}
\title{Annotate Cell Types using Custom GPT}
\arguments{
\item{input}{A list or data frame containing gene markers.}

\item{tissuename}{A character string specifying the tissue name.}

\item{model}{A character string specifying the model to use.}

\item{topgenenumber}{An integer specifying the number of top genes to consider.}

\item{baseurl}{A character string specifying the base URL for the API. Default is "https://api.default.com/v1/chat/completions".}
}
\value{
A character vector of annotated cell types.
}
\description{
This function annotates cell types based on input gene markers.
}
